% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aux_ZIBP_Laksh}
\alias{aux_ZIBP_Laksh}
\title{Auxiliar cpp function to obtain the probability for a vector x.}
\usage{
aux_ZIBP_Laksh(input)
}
\arguments{
\item{input}{numeric value with the information.}
}
\value{
returns the pmf for a vector.
}
\description{
Auxiliar cpp function to obtain the probability for a vector x.
}
