% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBP_Laksh.R
\name{correct_alpha_BP_Laksh}
\alias{correct_alpha_BP_Laksh}
\title{Correct value for alpha parameter - Lakshminarayana}
\usage{
correct_alpha_BP_Laksh(l1, l2)
}
\arguments{
\item{l1}{mean for the marginal \eqn{X_1} variable with Poisson distribution.}

\item{l2}{mean for the marginal \eqn{X_2} variable with Poisson distribution.}
}
\value{
Returns the density for a given data \code{x}.
}
\description{
This function obtains the interval in which the parameter \eqn{\alpha} must be.
}
\examples{
correct_alpha_BP_Laksh(l1=3, l2=4)
correct_alpha_BP_Laksh(l1=4, l2=3)
correct_alpha_BP_Laksh(l1=4, l2=4)

}
\author{
Freddy Hernandez-Barajas, \email{fhernanb@unal.edu.co}
}
