% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBP_Laksh.R
\name{dBP_Laksh}
\alias{dBP_Laksh}
\title{Bivariate Poisson Distribution - Laksh}
\usage{
dBP_Laksh(x, l1, l2, alpha, log = FALSE)
}
\arguments{
\item{x}{vector or matrix of quantiles. When \code{x} is a matrix,
each row is taken to be a quantile and columns correspond to the number of dimensions \code{p}.}

\item{l1}{mean for the marginal \eqn{X_1} variable with Poisson distribution.}

\item{l2}{mean for the marginal \eqn{X_2} variable with Poisson distribution.}

\item{alpha}{third parameter.}

\item{log}{logical; if \code{TRUE}, densities d are given as log(d).}
}
\value{
Returns the density for a given data \code{x}.
}
\description{
This function obtains the probability for the Bivariate Poisson distribution
under the parameterization of Lakshminarayana et. al (1993).
}
\examples{
# Probability for single values of X1 and X2
dBP_Laksh(c(0, 0), l1=3, l2=4, alpha=1)
dBP_Laksh(c(1, 0), l1=3, l2=4, alpha=1)
dBP_Laksh(c(0, 1), l1=3, l2=4, alpha=1)

# Probability for a matrix the values of X1 and X2
x <- matrix(c(0, 0,
              1, 0,
              0, 1), ncol=2, byrow=TRUE)
x
dBP_Laksh(x=x, l1=3, l2=4, alpha=1)

# Checking if the probabilites sum 1
val_x1 <- val_x2 <- 0:50
space <- expand.grid(val_x1, val_x2)
space <- as.matrix(space)

l1 <- 3
l2 <- 4
alpha <- -1.27

probs <- dBP_Laksh(x=space, l1=l1, l2=l2, alpha=alpha)
sum(probs)

}
\author{
Freddy Hernandez-Barajas, \email{fhernanb@unal.edu.co}
}
