% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dZIBP_Geoffroy.R
\name{dZIBP_Geoffroy}
\alias{dZIBP_Geoffroy}
\alias{rZIBP_Geoffroy}
\title{Zero Inflated Bivariate Poisson Distribution - Geoffroy}
\usage{
dZIBP_Geoffroy(x, l1, l2, l0, psi, log = FALSE)

rZIBP_Geoffroy(n, l1, l2, l0, psi)
}
\arguments{
\item{x}{vector or matrix of quantiles. When \code{x} is a matrix,
each row is taken to be a quantile and columns correspond to the number of dimensions \code{p}.}

\item{l1}{mean for \eqn{Z_1} variable with Poisson distribution.}

\item{l2}{mean for \eqn{Z_2} variable with Poisson distribution.}

\item{l0}{mean for \eqn{U} variable with Poisson distribution.}

\item{psi}{parameter with the contamination proportion, \eqn{0 \leq \psi \leq 1}.}

\item{log}{logical; if \code{TRUE}, densities d are given as log(d).}

\item{n}{number of random observations.}
}
\value{
Returns the density for a given data \code{x}.
}
\description{
This function obtains the probability for the Zero Inflated
Bivariate Poisson distribution
under the parameterization of Geoffroy et. al (2021).
}
\examples{

# Example 1 ---------------------------------------------------------------
# Probability for single values of X1 and X2
dZIBP_Geoffroy(c(0, 0), l1=3, l2=4, l0=1, psi=0.15)
dZIBP_Geoffroy(c(1, 0), l1=3, l2=4, l0=1, psi=0.15)
dZIBP_Geoffroy(c(0, 1), l1=3, l2=4, l0=1, psi=0.15)

# Probability for a matrix the values of X1 and X2
x <- matrix(c(0, 0,
              1, 0,
              0, 1), ncol=2, byrow=TRUE)
x
dZIBP_Geoffroy(x=x, l1=3, l2=4, l0=1, psi=0.15)

# Checking if the probabilities sum 1
val_x1 <- val_x2 <- 0:50
space <- expand.grid(val_x1, val_x2)
space <- as.matrix(space)

l1 <- 3
l2 <- 4
l0 <- 1.27
psi <- 0.15

probs <- dZIBP_Geoffroy(x=space, l1=l1, l2=l2, l0=l0, psi=psi)
sum(probs)

# Example 2 ---------------------------------------------------------------
# Heat map for a ZIBP_Geoffroy

l1 <- 3
l2 <- 4
l0 <- 1.2
psi <- 0.03

X1 <- 0:10
X2 <- 0:10
data <- expand.grid(X1=X1, X2=X2)
data$Prob <- dZIBP_Geoffroy(x=data, l1=l1, l2=l2, l0=l0, psi=psi)
data$X1 <- factor(data$X1)
data$X2 <- factor(data$X2)

library(ggplot2)
ggplot(data, aes(X1, X2, fill=Prob)) +
  geom_tile() +
  scale_fill_gradient(low="darkgreen", high="yellow")


# Example 3 ---------------------------------------------------------------
# Generating random values and moment estimations

l1 <- 13
l2 <- 8
l0 <- 5
psi <- 0.15

x <- rZIBP_Geoffroy(n=500, l1, l2, l0, psi)
moments_estim_ZIBP_Geoffroy(x)

# Example 4 ---------------------------------------------------------------
# Estimating the parameters using the loglik function

# Loglik function
llZIBP_Geoffroy <- function(param, x) {
  l1 <- param[1]  # param: is the parameter vector
  l2 <- param[2]
  l0 <- param[3]
  psi <- param[4]
  sum(dZIBP_Geoffroy(x=x, l1=l1, l2=l2,
                     l0=l0, psi=psi, log=TRUE))
}

# The known parameters
l1 <- 5
l2 <- 3
psi <- 0.20
l0 <- 1.20

set.seed(12345)
x <- rZIBP_Geoffroy(n=500, l1=l1, l2=l2, l0=l0, psi=psi)

# To obtain reasonable values for l0
start_param <- moments_estim_ZIBP_Geoffroy(x)
start_param

# Estimating parameters
res1 <- optim(fn = llZIBP_Geoffroy,
              par = start_param,
              lower = c(0.001, 0.001, 0.001, 0.0001),
              upper = c(  Inf,   Inf,   Inf, 0.9999),
              method = "L-BFGS-B",
              control = list(maxit=100000, fnscale=-1),
              x=x)

res1

}
\references{
Kouakou, K. J. G., Hili, O., & Dupuy, J. F. (2021). Estimation in the zero-inflated bivariate Poisson model with an application to health-care utilization data. Afrika Statistika, 16(2), 2767-2788.
}
\author{
Freddy Hernandez-Barajas, \email{fhernanb@unal.edu.co}
}
