% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marZIBP_HOL.R
\name{marZIBP_HOL}
\alias{marZIBP_HOL}
\title{Marginal Zero Inflated Bivariate Poisson Distribution - Holgate}
\usage{
marZIBP_HOL(
  mu1.fo,
  mu2.fo,
  psi.fo,
  data,
  initial.values = NULL,
  optimizer = "nlminb"
)
}
\arguments{
\item{mu1.fo}{a formula object to explain the marginal mean \eqn{\mu_1}, the first response \eqn{X_1} on the left of an ~ operator, and the terms, separated by operators, on the right.}

\item{mu2.fo}{a formula object to explain the marginal mean \eqn{\mu_2}, the second response \eqn{X_2} on the left of an ~ operator, and the terms, separated by operators, on the right.}

\item{psi.fo}{a formula object to explain the \eqn{\psi} parameter. An example could be: \code{~ age + salary}.}

\item{data}{a data frame containing the variables occurring in the formula2.}

\item{initial.values}{a vector with possible initial values for the parameters.}

\item{optimizer}{the optimizer to be used: nlminb, optim.}
}
\value{
Returns a object with class "marZIB_PHOL".
}
\description{
This function fit a marginal Zero Inflated Bivariate Poisson
distribution under the Holagate (1964) version.
}
\examples{
# Example 1 ---------------------------------------------------------------
l1 <- 3
l2 <- 4
l0 <- 1.5
psi <- 0.20

set.seed(12345678)
data1 <- rZIBP_HOL(n=500, l1=l1, l2=l2, l0=l0, psi=psi)
data1 <- as.data.frame(data1)

# To fit the model
mod1 <- NULL
mod1 <- marZIBP_HOL(mu1.fo=X1~1,
                    mu2.fo=X2~1,
                    psi.fo=~1,
                    data=data1)

# To obtain the usual summary table
summary(mod1)

# To explore the estimations of l1, l2, mu and p

# To obtain E(Y1)=v1 and E(Y2)=v2
mod1$fitted.mu1[1]
mod1$fitted.mu2[1]

# To compare sample means with v1 and v2
colMeans(data1)

# To obtain l0 and psi
mod1$fitted.l0
mod1$fitted.psi[1]

# To obtain l1 and l2
mod1$fitted.l1[1]
mod1$fitted.l2[1]

# Example 2 ---------------------------------------------------------------

gen_data_ZIBP_HOL <- function(n=100) {
  # To generate the covariates
  x1 <- runif(n=n)
  x2 <- runif(n=n)

  # To generate the means
  mu1 <- exp(1 + 1.3 * x1)
  mu2 <- exp(1 + 2.1 * x2)

  # To generate the psi
  logit_inv <- function(x) exp(x) / (1+exp(x))
  psi <- logit_inv(-2.4 + 2.1 * x2)

  # The third lambda
  l0 <- 1

  # To obtain lambdas
  l1 <- mu1 / (1-psi) - l0
  l2 <- mu2 / (1-psi) - l0

  # To generate Y1 and Y2
  y <- NULL
  for (i in 1:n)
    y <- rbind(y, rZIBP_HOL(n=1, l1=l1[i], l2=l2[i],
                                 l0=l0, psi=psi[i]))

  # To create the dataset
  dataset <- data.frame(y1=y[,1], y2=y[,2],
                        x1=x1, x2=x2,
                        mu1=mu1, mu2=mu2,
                        l0=l0, psi=psi,
                        l1=l1, l2=l2)

  return(dataset)
}

set.seed(12)
data2 <- gen_data_ZIBP_HOL(n=300)
head(data2, n=8)

mod2 <- NULL
mod2 <- marZIBP_HOL(mu1.fo=y1~x1,
                    mu2.fo=y2~x2,
                    psi.fo=~x2,
                    data=data2)

summary(mod2)


}
\references{
P. Holgate. Estimation for the bivariate poisson distribution.
Biometrika, 51(1/2):241–245, 1964. ISSN 00063444, 14643510.
URL http://www.jstor.org/stable/2334210.

Konan Jean Geoffroy Kouakou, Ouagnina Hili, and Jean-Francois
Dupuy. Estimation in the zero-inflated bivariate poisson model
with an application to health-care utilization data.
Afrika Statistika, 16(2):2767–2788, 2021.
}
\author{
Freddy Hernandez-Barajas, \email{fhernanb@unal.edu.co}
}
