% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBP_Laksh.R
\name{moments_estim_BP_Laksh}
\alias{moments_estim_BP_Laksh}
\title{Moment estimations for Bivariate Poisson Distribution - Laksh}
\usage{
moments_estim_BP_Laksh(x)
}
\arguments{
\item{x}{vector or matrix of quantiles. When \code{x} is a matrix,
each row is taken to be a quantile and columns correspond to the number of dimensions \code{p}.}
}
\value{
Returns a vector with \eqn{\hat{\lambda_1}}, \eqn{\hat{\lambda_2}} and \eqn{\hat{\alpha}}.
}
\description{
This function obtains moment estimators for the Bivariate Poisson distribution
under the parameterization of Lakshminarayana et. al (1993).
}
\examples{
# Generating random values and moment estimations
l1 <- 1
l2 <- 2
alpha <- -2.7
x <- rBP_Laksh(n=500, l1, l2, alpha)

moments_estim_BP_Laksh(x)

# Analizing example from Famoye (2010)

freq <- c(34, 20, 4, 6, 4,
          17, 7, 0, 0, 0,
          6, 4, 1, 0, 0,
          0, 4, 0, 0, 0,
          0, 0, 0, 0, 0,
          2, 0, 0, 0, 0)

data_table <- matrix(freq, ncol=5, byrow=TRUE)
rownames(data_table) <- 0:5
colnames(data_table) <- 0:4

data_table

long <- as.data.frame.table(data_table)
x <- long[rep(1:nrow(long), long$Freq), -3]
x <- data.matrix(x)

colMeans(x)
var(x)
cor(x)
moments_estim_BP_Laksh(x)
}
\references{
Lakshminarayana, J., Pandit, S. N., & Srinivasa Rao, K. (1999). On a bivariate Poisson distribution. Communications in Statistics-Theory and Methods, 28(2), 267-276.
}
\author{
Freddy Hernandez-Barajas, \email{fhernanb@unal.edu.co}
}
