% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dZIBP_Laksh.R
\name{moments_estim_ZIBP_Laksh}
\alias{moments_estim_ZIBP_Laksh}
\title{Moment estimations for Zero Inflated Bivariate Poisson Distribution - Laksh}
\usage{
moments_estim_ZIBP_Laksh(x)
}
\arguments{
\item{x}{vector or matrix of quantiles. When \code{x} is a matrix,
each row is taken to be a quantile and columns correspond to the number of dimensions \code{p}.}
}
\value{
Returns a vector with \eqn{\hat{\lambda_1}}, \eqn{\hat{\lambda_2}}, \eqn{\hat{\alpha}} and \eqn{\hat{\psi}}.
}
\description{
This function obtains moment estimators for the Zero Inflated
Bivariate Poisson distribution
under the parameterization of Lakshminarayana et. al (1993).
}
\examples{
# Generating random values and moment estimations
l1 <- 1
l2 <- 2
alpha <- -2.7
psi <- 0.15
x <- rZIBP_Laksh(n=500, l1, l2, alpha, psi)

moments_estim_ZIBP_Laksh(x)
}
\references{
Lakshminarayana, J., Pandit, S. N., & Srinivasa Rao, K. (1999). On a bivariate Poisson distribution. Communications in Statistics-Theory and Methods, 28(2), 267-276.
}
\author{
Freddy Hernandez-Barajas, \email{fhernanb@unal.edu.co}
}
