% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIBP_Geoffroy.R, R/marZIBP_Geoffroy.R
\name{print.ZIBPGeoffroy}
\alias{print.ZIBPGeoffroy}
\alias{summary.marZIBPGeoffroy}
\alias{print.marZIBPGeoffroy}
\title{Summary table for marZIP - Geoffroy}
\usage{
\method{print}{ZIBPGeoffroy}(object, ...)

\method{summary}{marZIBPGeoffroy}(object, ...)

\method{print}{marZIBPGeoffroy}(object, ...)
}
\arguments{
\item{object}{of class marZIBPGeoffroy.}

\item{...}{aditional arguments.}
}
\value{
Returns the summary table.
}
\description{
This function obtains the summary table for objects of class marZIBPGeoffroy.
}
\examples{
# Example 1 ---------------------------------------------------------------
l1 <- 3
l2 <- 4
l0 <- 1.5
psi <- 0.20

set.seed(12345678)
data1 <- rZIBP_Geoffroy(n=500, l1=l1, l2=l2, l0=l0, psi=psi)
data1 <- as.data.frame(data1)

# To fit the model
mod1 <- NULL
mod1 <- marZIBP_Geoffroy(mu1.fo=X1~1,
                         mu2.fo=X2~1,
                         psi.fo=~1,
                         data=data1)

# To obtain the usual summary table
summary(mod1)

# To explore the estimations of l1, l2, mu and p

# To obtain E(Y1)=v1 and E(Y2)=v2
mod1$fitted.mu1[1]
mod1$fitted.mu2[1]

# To compare sample means with v1 and v2
colMeans(data1)

# To obtain l0 and psi
mod1$fitted.l0
mod1$fitted.psi[1]

# To obtain l1 and l2
mod1$fitted.l1[1]
mod1$fitted.l2[1]

# Example 2 ---------------------------------------------------------------

gen_data_ZIBP_Geoffroy <- function(n=100) {
  # To generate the covariates
  x1 <- runif(n=n)
  x2 <- runif(n=n)

  # To generate the means
  mu1 <- exp(1 + 1.3 * x1)
  mu2 <- exp(1 + 2.1 * x2)

  # To generate the psi
  logit_inv <- function(x) exp(x) / (1+exp(x))
  psi <- logit_inv(-2.4 + 2.1 * x2)

  # The third lambda
  l0 <- 1

  # To obtain lambdas
  l1 <- mu1 / (1-psi) - l0
  l2 <- mu2 / (1-psi) - l0

  # To generate Y1 and Y2
  y <- NULL
  for (i in 1:n)
    y <- rbind(y, rZIBP_Geoffroy(n=1, l1=l1[i], l2=l2[i],
                                 l0=l0, psi=psi[i]))

  # To create the dataset
  dataset <- data.frame(y1=y[,1], y2=y[,2],
                        x1=x1, x2=x2,
                        mu1=mu1, mu2=mu2,
                        l0=l0, psi=psi,
                        l1=l1, l2=l2)

  return(dataset)
}

set.seed(12)
data2 <- gen_data_ZIBP_Geoffroy(n=300)
head(data2, n=8)

mod2 <- NULL
mod2 <- marZIBP_Geoffroy(mu1.fo=y1~x1,
                         mu2.fo=y2~x2,
                         psi.fo=~x2,
                         data=data2)

summary(mod2)


}
\author{
Freddy Hernandez-Barajas, \email{fhernanb@unal.edu.co}
}
