% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBP_Laksh.R
\name{dBP_Laksh}
\alias{dBP_Laksh}
\alias{rBP_Laksh}
\title{Bivariate Poisson Distribution - Laksh}
\usage{
dBP_Laksh(x, l1, l2, alpha, log = FALSE)

rBP_Laksh(n, l1, l2, alpha, max_val_x1 = NULL, max_val_x2 = NULL)
}
\arguments{
\item{x}{vector or matrix of quantiles. When \code{x} is a matrix,
each row is taken to be a quantile and columns correspond to the number of dimensions \code{p}.}

\item{l1}{mean for the marginal \eqn{X_1} variable with Poisson distribution.}

\item{l2}{mean for the marginal \eqn{X_2} variable with Poisson distribution.}

\item{alpha}{third parameter.}

\item{log}{logical; if \code{TRUE}, densities d are given as log(d).}

\item{n}{number of random observations.}

\item{max_val_x1}{maximum value for \eqn{X_1} that is expected, by default it is 100.}

\item{max_val_x2}{maximum value for \eqn{X_2} that is expected, by default it is 100.}
}
\value{
Returns the density for a given data \code{x}.
}
\description{
This function obtains the probability for the Bivariate Poisson distribution
under the parameterization of Lakshminarayana et. al (1993).
}
\examples{
# Probability for single values of X1 and X2
dBP_Laksh(c(0, 0), l1=3, l2=4, alpha=1)
dBP_Laksh(c(1, 0), l1=3, l2=4, alpha=1)
dBP_Laksh(c(0, 1), l1=3, l2=4, alpha=1)

# Probability for a matrix the values of X1 and X2
x <- matrix(c(0, 0,
              1, 0,
              0, 1), ncol=2, byrow=TRUE)
x
dBP_Laksh(x=x, l1=3, l2=4, alpha=1)

# Checking if the probabilities sum 1
val_x1 <- val_x2 <- 0:50
space <- expand.grid(val_x1, val_x2)
space <- as.matrix(space)

l1 <- 3
l2 <- 4
alpha <- -1.27

probs <- dBP_Laksh(x=space, l1=l1, l2=l2, alpha=alpha)
sum(probs)

# Heat map for a BP_Laksh

l1 <- 1
l2 <- 2
correct_alpha_BP_Laksh(l1=l1, l2=l2)
alpha <- -2.9

X1 <- 0:10
X2 <- 0:10
data <- expand.grid(X1=X1, X2=X2)
data$Prob <- dBP_Laksh(x=data, l1=l1, l2=l2, alpha=alpha)
data$X1 <- factor(data$X1)
data$X2 <- factor(data$X2)

library(ggplot2)
ggplot(data, aes(X1, X2, fill=Prob)) +
  geom_tile() +
  scale_fill_gradient(low="darkgreen", high="white")

# Generating random values and moment estimations
l1 <- 1
l2 <- 2
correct_alpha_BP_Laksh(l1=l1, l2=l2)
alpha <- -2.7

x <- rBP_Laksh(n=500, l1, l2, alpha)
moments_estim_BP_Laksh(x)
}
\references{
Lakshminarayana, J., Pandit, S. N., & Srinivasa Rao, K. (1999). On a bivariate Poisson distribution. Communications in Statistics-Theory and Methods, 28(2), 267-276.
}
\author{
Freddy Hernandez-Barajas, \email{fhernanb@unal.edu.co}
}
