% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBP_Laksh.R
\name{dBP_Laksh}
\alias{dBP_Laksh}
\title{Bivariate Poisson Distribution - Lakshminarayana}
\usage{
dBP_Laksh(x, l1, l2, alpha, log = FALSE)
}
\arguments{
\item{x}{vector or matrix of quantiles. When \code{x} is a matrix, each row is taken to be a quantile and columns correspond to the number of dimensions \code{p}.}

\item{l1}{mean for the marginal \eqn{X_1} variable with Poisson distribution.}

\item{l2}{mean for the marginal \eqn{X_2} variable with Poisson distribution.}

\item{alpha}{third parameter.}

\item{log}{logical; if \code{TRUE}, densities d are given as log(d).}
}
\value{
Returns the density for a given data \code{x}.
}
\description{
Bivariate Poisson Distribution - Lakshminarayana
}
\examples{
# Pruebas para chequear vectorizacion de la funcion de densidad
dBP_Laksh(c(0, 0), l1=3, l2=4, alpha=1)
dBP_Laksh(c(1, 0), l1=3, l2=4, alpha=1)
dBP_Laksh(c(0, 1), l1=3, l2=4, alpha=1)

muestra <- matrix(c(0, 0,
                    1, 0,
                    0, 1), ncol=2, byrow=TRUE)
muestra
dBP_Laksh(x=muestra, l1=3, l2=4, alpha=1)

}
