% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBP_Laksh.R
\name{dBP_Laksh}
\alias{dBP_Laksh}
\alias{rBP_Laksh}
\title{Bivariate Poisson Distribution - Laksh}
\usage{
dBP_Laksh(x, l1, l2, alpha, log = FALSE)

rBP_Laksh(n, l1, l2, alpha, max_val_x1 = NULL, max_val_x2 = NULL)
}
\arguments{
\item{x}{vector or matrix of quantiles. When \code{x} is a matrix,
each row is taken to be a quantile and columns correspond to the number of dimensions \code{p}.}

\item{l1}{mean for the marginal \eqn{X_1} variable with Poisson distribution.}

\item{l2}{mean for the marginal \eqn{X_2} variable with Poisson distribution.}

\item{alpha}{third parameter.}

\item{log}{logical; if \code{TRUE}, densities d are given as log(d).}

\item{n}{number of random observations.}

\item{max_val_x1}{maximum value for \eqn{X_1} that is expected, by default it is 100.}

\item{max_val_x2}{maximum value for \eqn{X_2} that is expected, by default it is 100.}
}
\value{
Returns the density for a given data \code{x}.
}
\description{
This function obtains the probability for the Bivariate Poisson distribution
under the parameterization of Lakshminarayana et. al (1993).
}
\examples{

# Example 1 ---------------------------------------------------------------
# Probability for single values of X1 and X2
dBP_Laksh(c(0, 0), l1=3, l2=4, alpha=1)
dBP_Laksh(c(1, 0), l1=3, l2=4, alpha=1)
dBP_Laksh(c(0, 1), l1=3, l2=4, alpha=1)

# Probability for a matrix the values of X1 and X2
x <- matrix(c(0, 0,
              1, 0,
              0, 1), ncol=2, byrow=TRUE)
x
dBP_Laksh(x=x, l1=3, l2=4, alpha=1)

# Checking if the probabilities sum 1
val_x1 <- val_x2 <- 0:50
space <- expand.grid(val_x1, val_x2)
space <- as.matrix(space)

l1 <- 3
l2 <- 4
alpha <- -1.27

probs <- dBP_Laksh(x=space, l1=l1, l2=l2, alpha=alpha)
sum(probs)


# Example 2 ---------------------------------------------------------------
# Heat map for a BP_Laksh

l1 <- 1
l2 <- 2
correct_alpha_BP_Laksh(l1=l1, l2=l2)
alpha <- -2.9

X1 <- 0:10
X2 <- 0:10
data <- expand.grid(X1=X1, X2=X2)
data$Prob <- dBP_Laksh(x=data, l1=l1, l2=l2, alpha=alpha)
data$X1 <- factor(data$X1)
data$X2 <- factor(data$X2)

library(ggplot2)
ggplot(data, aes(X1, X2, fill=Prob)) +
  geom_tile() +
  scale_fill_gradient(low="darkgreen", high="white")


# Example 3 ---------------------------------------------------------------
# Generating random values and moment estimations
l1 <- 1
l2 <- 2
correct_alpha_BP_Laksh(l1=l1, l2=l2)
alpha <- -2.7

x <- rBP_Laksh(n=500, l1, l2, alpha)
moments_estim_BP_Laksh(x)


# Example 4 ---------------------------------------------------------------
# Estimating the parameters using the loglik function

# Loglik function
llBP_Laksh <- function(param, x) {
  l1    <- param[1]  # param: is the parameter vector
  l2    <- param[2]
  alpha <- param[3]
  sum(dBP_Laksh(x=x, l1=l1, l2=l2, alpha=alpha, log=TRUE))
}



l1 <- 1
l2 <- 2

correct_alpha_BP_Laksh(l1=l1, l2=l2)

alpha <- -2.7

set.seed(12345)
x <- rBP_Laksh(n=500, l1=l1, l2=l2, alpha=alpha)

# To obtain reasonable values for alpha
theta <- as.numeric(moments_estim_BP_Laksh(x))
theta

# To create start parameters
min_alpha <- correct_alpha_BP_Laksh(l1=theta[1],
                                    l2=theta[2])$min_alpha
max_alpha <- correct_alpha_BP_Laksh(l1=theta[1],
                                    l2=theta[2])$max_alpha

start_param <- theta
names(start_param) <- c("l1_hat", "l2_hat", "alpha_hat")
start_param

# Estimating parameters
res1 <- optim(fn = llBP_Laksh,
              par = start_param,
              lower = c(0.001, 0.001, min_alpha),
              upper = c(  Inf,   Inf, max_alpha),
              method = "L-BFGS-B",
              control = list(maxit=100000, fnscale=-1),
              x=x)

res1

# Analizing example of Laksh (1999) ---------------------------------------

x <- rep(0:5, each=6)
y <- rep(0:5, times=6)
freq <- c(7,  41, 54, 40, 21, 9,
          36, 79, 83, 59, 30, 13,
          39, 70, 69, 47, 25, 10,
          24, 41, 39, 26, 14, 6,
          10, 18, 18, 11, 6, 2,
          3, 6, 6, 4, 2, 1)

seed_plants <- NULL
for (i in 1:36) {
  temp <- matrix(c(x[i], y[i]),  ncol=2, nrow=freq[i], byrow=TRUE)
  seed_plants <- rbind(seed_plants, temp)
}

seed_plants

# Exploring some statistics
colMeans(seed_plants)
var(seed_plants)
cor(seed_plants)

# Moment estimators
moments_estim_BP_Laksh(seed_plants)

# Correct interval for alpha according to l1_hat and l2_hat
correct_alpha_BP_Laksh(l1=1.6925, l2=2.0134)

# Finding the log-likelihood estimators using
optim(fn = llBP_Laksh,
      par = c(1.6925, 2.0134, -1.3230),
      lower = c(0.0001, 0.0001, -2.114373),
      upper = c(Inf, Inf, 2.114373),
      method = "L-BFGS-B",
      control = list(maxit=100000, fnscale=-1),
      x=seed_plants)
}
\references{
Lakshminarayana, J., Pandit, S. N., & Srinivasa Rao, K. (1999). On a bivariate Poisson distribution. Communications in Statistics-Theory and Methods, 28(2), 267-276.
}
\author{
Freddy Hernandez-Barajas, \email{fhernanb@unal.edu.co}
}
